Move samples to /samples dir, refactor for commonjs
